Actualizar el contenido de la rama my_uname
